requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-socketio==5.10.0
python-engineio==4.8.0
simple-websocket==1.0.0
